[UPDATE, FIX] Implement grantAdmin API and UPDATE delete user after logout

+++ server/iam/src/main/java/com/lhsk/iam/domain/admin/service/AdminService.java
@@ -4,6 +4,9 @@ import java.security.GeneralSecurityException;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
@@ -13,8 +16,14 @@ import com.lhsk.iam.domain.account.model.vo.UserAccountVO;
 import com.lhsk.iam.domain.account.service.AccountService;
 import com.lhsk.iam.domain.admin.model.mapper.AdminMapper;
 import com.lhsk.iam.domain.admin.model.vo.DeptVO;
+import com.lhsk.iam.domain.admin.model.vo.MenuClickRequestVO;
 import com.lhsk.iam.domain.admin.model.vo.MenuClickVO;
 import com.lhsk.iam.domain.user.model.vo.LoginHistoryVO;
+import com.lhsk.iam.domain.user.model.vo.UpdateUserVO;
+import com.lhsk.iam.domain.user.service.LoginService;
+import com.lhsk.iam.domain.user.service.UserService;
+import com.lhsk.iam.global.config.JwtConfig;
+import com.lhsk.iam.global.config.jwt.JwtTokenProvider;
 import com.lhsk.iam.global.encrypt.AesGcmEncrypt;
 
 import lombok.RequiredArgsConstructor;
@@ -27,6 +36,11 @@ public class AdminService {
 	
 	private final AdminMapper adminMapper;
 	private final AccountService accountService;
+	private final UserService userService;
+	private final LoginService loginService;
+	private final JwtTokenProvider jwtTokenProvider;
+	private final JwtConfig jwtConfig;
+	
 	private AesGcmEncrypt aesGcmEncrypt = new AesGcmEncrypt();
 	
 	@Value("${aes.secret}")
@@ -36,6 +50,16 @@ public class AdminService {
 	// 최종적으로 iv가 저장될 byte[] 변수
 	private byte[] iv = new byte[12];
 	
+	// iv property를 byte[]로 변환
+	public byte[] ivToByteArray(String ivString) {
+		// property에서 String으로 받아온 ivString을  ", "을 기준으로 split -> String[]에 저장
+		String[] ivStringArray = ivString.split(", ");
+		// String[] -> byte[]로 번환
+		for (int i = 0; i < iv.length; i++) {
+			iv[i] = Byte.parseByte(ivStringArray[i]);
+		}
+		return iv;
+	}
 	
 	// 모든 유저의 로그인기록 조회
 	public List<LoginHistoryVO> findAllLoginHistory() {
@@ -155,20 +179,24 @@ public class AdminService {
 	public void grantAvailableAccounts(int userNo, List<UserAccountVO> data) {
 		// iv 값 세팅
 		ivToByteArray(ivString);
-		// 클라이언트에서 넘어 온 값 재가공
-		for (UserAccountVO info : data) {
-			try {
-				info.setUserNo(userNo);
-				// 계좌번호 암호화
-				info.setAcctNo(aesGcmEncrypt.encrypt(info.getAcctNo(), key, iv));
-			} catch (GeneralSecurityException e) {
-				e.printStackTrace();
+		// data가 null이면 delete만 수행(해당 회원의 허용 계좌 전부 삭제)
+		if (data == null) adminMapper.deleteUserAccount(userNo);
+		else {			
+			// 클라이언트에서 넘어 온 값 재가공
+			for (UserAccountVO info : data) {
+				try {
+					info.setUserNo(userNo);
+					// 계좌번호 암호화
+					info.setAcctNo(aesGcmEncrypt.encrypt(info.getAcctNo(), key, iv));
+				} catch (GeneralSecurityException e) {
+					e.printStackTrace();
+				}
 			}
+			int deleteCount = adminMapper.deleteUserAccount(userNo);
+			int insertCount = adminMapper.insertUserAccount(data);
+			log.info("deleteCount: "+ deleteCount+", insertCount: "+insertCount);
 		}
-		int deleteCount = adminMapper.deleteUserAccount(userNo);
-		int insertCount = adminMapper.insertUserAccount(data);
 		
-		log.info("delete: "+ deleteCount+", insertCount: "+insertCount);
 	}
 	
 	// 부서 조회
@@ -209,16 +237,38 @@ public class AdminService {
 		return false;
 	}
 
-	// iv property를 byte[]로 변환
-	public byte[] ivToByteArray(String ivString) {
-		// property에서 String으로 받아온 ivString을  ", "을 기준으로 split -> String[]에 저장
-		String[] ivStringArray = ivString.split(", ");
-		// String[] -> byte[]로 번환
-		for (int i = 0; i < iv.length; i++) {
-		    iv[i] = Byte.parseByte(ivStringArray[i]);
-		}
-		return iv;
+	// 권한 위임
+	public boolean grantAdmin(HttpServletRequest request, HttpServletResponse response, UpdateUserVO updateUserVO) {
+		String accessToken = request.getHeader("Authorization")
+						.replace(jwtConfig.getTokenPrefix(), "");
+		int adminsUserNo = jwtTokenProvider.getUserNoFromToken(accessToken);
+		try {
+			// 위임 받는 매니저의 정보 업데이트
+			userService.updateUser("ROLE_ADMIN", updateUserVO);			
+			// 위임 후 ROLE_ADMIN을 ROLE_USER로 바꿈
+			adminMapper.afterGrantAdmin(adminsUserNo);
+			// 로그아웃
+			// 메뉴 클릭 기록 집계 등록 (insert or update)
+			MenuClickRequestVO vo = MenuClickRequestVO
+								.builder()
+								.allAccount(0)
+								.inout(0)
+								.inoutReport(0)
+								.dailyReport(0)
+								.dashboard(0)
+								.build();
+			userService.updateMenuClick(vo);
+			// 로그아웃 처리
+			loginService.logout(request, response);
+			return true;
+		} catch (Exception e) {
+			e.printStackTrace();
+		} 
+		return false;
 	}
+
+	
+
